+++ README.md
@@ -1,10 +1,11 @@
 # PORTFOLIO
 
-총 5페이지로 이루어진 포트폴리오 입니다: HOME,THREEJS,PROJECT,BOOK,CONTACT
+총 5페이지로 이루어진 About Me 포트폴리오 입니다: HOME,THREEJS,PROJECT,BOOK,CONTACT
 
 ## ARCHTECTURE
 
-│ ├─ └─ 📃 📂
+- React + Typescript 기반의 포트폴리오 웹 사이트
+- Atomic Design 패턴을 적용한 컴포넌트 구조
 
 ```
 SELFBLEY
@@ -16,100 +17,63 @@ SELFBLEY
 │ │ │ ├─📂Text
 │ │ │ │ └─📃index.tsx
 │ │ │ └─ ...
-│ │ ├─📂loading # WAIT
-│ │ ├─📂molecules # atoms + 기능/UI 를 조합한 하나의 기능/UI 컴포넌트
+│ │ ├─📂molecules # 여러 atoms을 조합한 하나의 기능 또는 UI 블록 컴포넌트
 │ │ │ ├─📂 FixedButton # Button + UI
 │ │ │ │ └─📃index.tsx
 │ │ │ ├─📂 Marquee # Text + Animation
 │ │ │ │ └─📃index.tsx
-│ │ │ ├─📂 Nav # Text + navigation
-│ │ │ │ └─📃index.tsx
+│ │ │ ├─📂 BookCard # Text + navigation
+│ │ │ │ └─📃index.**tsx**
 │ │ │ └─ ...
-│ │ ├─📂organism
-│ │ │ ├─📂 FixedButton # Button + UI
-│ │ │ │ └─📃index.tsx
-│ │ │ ├─📂 Marquee # Text + Animation
-│ │ │ │ └─📃index.tsx
-│ │ │ ├─📂 Nav # Text + navigation
-│ │ │ │ └─📃index.tsx
+│ │ ├─📂organism # 여러 molecules / atoms을 조합해 복합적인 UI 섹션을 구성하는 단위
+│ │ │ ├─📂 BookSection
+│ │ │ │ ├─📃BookBackground.tsx # BookSection 배경 연출 및 스크롤에 따른 시각 효과 담당
+│ │ │ │ ├─📃BookList.tsx # 책 데이터를 기반으로 카드 리스트를 렌더링
+│ │ │ │ └─📃index.tsx # 책 섹션 레이아웃 및 스크롤 애니메이션 제어
 │ │ │ └─ ...
-│ │ ├─📂template
+│ │ ├─📂template # 페이지의 레이아웃 구조
+│ │ │ ├─📂container
+│ │ │ │ ├─📂SVGContainer
+│ │ │ │ │ └─📃index.tsx
+│ │ │ │ └─ ...
+│ │ │ └─ ...
 │ │ └─📂transition
+│ │   ├─📂FlipTransition
+│ │   │ └─📃index.tsx
+│ │   └─ ...
 │ ├─📂contants
 │ │ └─ ...
 │ ├─📂hooks
 │ │ └─ ...
 │ ├─📂layout
-│ │ ├─📂background
-│ │ ├─📂Footer
-│ │ └─📂Header
-│ ├─📂pages
 │ │ └─ ...
-│ ├─📂stores
+│ ├─📂pages # 라우트별 페이지 컴포넌트
+│ │ └─ ...
+│ ├─📂stores # Zustand 전역 상태 관리
+│ │ ├─📃useScreenStore.ts # 브라우저 반응형 상태
 │ │ └─ ...
 │ ├─📂styles
 │ │ └─ ...
 │ ├─📂types
 │ │ └─ ...
-│ ├─📃App.tsx
+│ ├─📃App.tsx # React Route
 │ └─ ...
 └─📃README.md
-
 ```
 
-```
-new_invitation/
-├─ README.md
-├─ package.json
-├─ public/
-│ └─ ...
-├─ src/
-│ ├─ app/
-│ │ ├─ api/
-│ │ │ └─ ... # route handlers
-│ │ ├─ profile/
-│ │ │ └─ page.tsx
-│ │ ├─ layout.tsx
-│ │ └─ ... # other Next.js pages/segments
-│ ├─ actions/ # Server Actions ("use server")
-│ │ └─ ...
-│ ├─ components/
-│ │ ├─ atoms/ # 가장 작은 단위의 컴포넌트
-│ │ │ ├─ Btn # 버튼 UI
-│ │ │ │ └─ index.tsx
-│ │ │ ├─ Icon # 아이콘 UI
-│ │ │ │ └─ index.tsx
-│ │ │ └─ ...
-│ │ ├─ molecules/ # 2개 이상의 Atom 으로 구성된 UI 컴포넌트
-│ │ │ ├─ btns/
-│ │ │ │ ├─ MusicBtn
-│ │ │ │ │ └─ index.tsx # 버튼 + Icon + 이벤트 핸들러
-│ │ │ │ └─ ...
-│ │ │ │
-│ │ │ └─ ...
-│ │ ├─ organisms/ # 페이지 내 블록 단위 컴포넌트
-│ │ │ ├─ panel/
-│ │ │ │ ├─ ParentInfoPanel
-│ │ │ │ │ ├─ constant.ts # 해당 컴포넌트 전용 상수 모음
-│ │ │ │ │ ├─ type.d.ts # 해당 컴포넌트 전용 타입 모음
-│ │ │ │ │ ├─ variants.ts # 해당 컴포넌트 전용 motion variants
-│ │ │ │ │ └─ index.tsx # Input + Btn 등 블록 단위 컴포넌트
-│ │ │ │ └─ ...
-│ │ │ │
-│ │ │ └─ ...
-│ │ └─ layout/ # Header, Footer 등 레이아웃 관련 컴포넌트
-│ ├─ hooks/ # React 훅 (클라이언트 전용 로직)
-│ ├─ store/ # zustand 등 전역 상태관리
-│ ├─ services/ # 외부 API / 비즈니스 로직 (user, invitation 등)
-│ ├─ lib/ # jose, cloudinary 설정 등 외부 라이브러리 설정/유틸 모음
-│ ├─ utils/ # 유틸 함수, validation 등
-│ ├─ types/ # 전역 타입 정의
-│ ├─ constants/ # 전역 상수
-│ └─ styles/ # 전역 스타일, motion variants (공용)
-│
-└─ ...
+## 외부 데이터 연동
 
-```
+### 1. My JSON Server (Serverless)
+
+- GitHub의 `db.json` 파일을 REST API로 제공
+- 엔드포인트: `/projects`
+- 프로젝트 목록 및 상세 정보 제공
+
+### 2. Naver Book API (Proxy)
+
+- **보안**: API 키 등 민감 정보의 클라이언트 노출 방지
+- **구현**: GitHub 서버리스 함수를 통한 프록시 요청
+- **사용법**: `/api/search?q={keyword}` → Naver Book API 호출 → 결과 반환
 
 ## HOMEPAGE
 
@@ -160,27 +124,3 @@ new_invitation/
 - Framer Motion
 - React
 - Zustand
-
-## Description
-
-2D 디자인을 바탕으로 정적인 느낌을 받지 않도록 하기 위해 다양한 애니메이션들을 넣어서 반응형 포트폴리오로 준비하게 됐습니다.
-
-Marquee : 중점적으로 생각하는 코드 스타일 키워드
-Scratch : 하고 싶은 말
-Parallax : 개발 일대기
-
-## Image 최적화
-
-Book SLider Section
-
-- BackgroundImg
-  - 사이즈: 최대 500kb 이하
-  - 확장자: .webp
-- Thumnail
-  - 사이즈: 최대 110kb 이하
-  - 확장자: .webp
-
-Parallax Section
-
-- BackgroundImg
-  - 사이즈: 최대 500kb 이하하
