+++ src/constants/technology/index.ts
@@ -1,0 +1,10 @@
+const language = ["html", "css", "javascript", "typescript"];
+const frontend = ["styledComponents", "react", "nextJs", "redux", "zustand"];
+const backend = ["nodeJs", "nextJs", "express", "mongodb", "firebase"];
+const etc = ["discord", "figma", "github"];
+export const technologys = {
+  language,
+  frontend,
+  backend,
+  etc,
+} as const;
